+++ ogn/parser/parse_skylines.py
@@ -1,0 +1,2 @@
+def parse(aprs_comment):
+    raise NotImplementedError("Skylines beacon parser not yet implemented")